Refactor E2E tests (#13244)
* Reuse default address in warehouse create util

* Drop unused parmeter from E2eApiClient

* Add assert for chekcoutId in created order via CheckoutCreate

* Add short description to steps

* Split Unit and E2E test in pytest workflow
