+++ saleor/tests/e2e/__init__.py
@@ -1,0 +1,12 @@
+DEFAULT_ADDRESS = {
+    "firstName": "John Saleor",
+    "lastName": "Doe Mirumee",
+    "companyName": "Saleor Commerce",
+    "streetAddress1": "14208 Hawthorne Blvd",
+    "streetAddress2": "",
+    "postalCode": "90250",
+    "country": "US",
+    "city": "Hawthorne",
+    "countryArea": "CA",
+    "phone": "+12025550163",
+}
