+++ saleor/tests/e2e/checkout/test_unlogged_customer_should_be_able_to_order_physical_product.py
@@ -122,7 +122,7 @@ def test_process_checkout_with_physical_product(
         channel_slug,
     )
 
-    # Step 1
+    # Step 1 - Create checkout.
     lines = [
         {"variantId": product_variant_id, "quantity": 1},
     ]
@@ -140,14 +140,14 @@ def test_process_checkout_with_physical_product(
     assert checkout_data["deliveryMethod"] is None
     assert checkout_data["shippingMethod"] is None
 
-    # Step 2
+    # Step 2 - Set shipping address for checkout.
     checkout_data = checkout_shipping_address_update(
         e2e_not_logged_api_client, checkout_id
     )
     assert len(checkout_data["shippingMethods"]) == 1
     shipping_method_id = checkout_data["shippingMethods"][0]["id"]
 
-    # Step 3
+    # Step 3 - Set DeliveryMethod for checkout.
     checkout_data = checkout_delivery_method_update(
         e2e_not_logged_api_client,
         checkout_id,
@@ -156,12 +156,12 @@ def test_process_checkout_with_physical_product(
     assert checkout_data["deliveryMethod"]["id"] == shipping_method_id
     total_gross_amount = checkout_data["totalPrice"]["gross"]["amount"]
 
-    # Step 4
+    # Step 4 - Create payment for checkout.
     checkout_dummy_payment_create(
         e2e_not_logged_api_client, checkout_id, total_gross_amount
     )
 
-    # Step 5
+    # Step 5 - Complete checkout.
     order_data = checkout_complete(e2e_not_logged_api_client, checkout_id)
     assert order_data["isShippingRequired"] is True
     assert order_data["status"] == "UNFULFILLED"
