+++ saleor/tests/e2e/checkout/utils/checkout_shipping_address_update.py
@@ -1,5 +1,5 @@
+from ... import DEFAULT_ADDRESS
 from ...utils import get_graphql_content
-from .. import DEFAULT_ADDRESS
 
 CHECKOUT_SHIPPING_ADDRESS_UPDATE_MUTATION = """
 mutation CheckoutShippingAddressUpdate(
