+++ saleor/tests/e2e/warehouse/utils.py
@@ -1,3 +1,4 @@
+from .. import DEFAULT_ADDRESS
 from ..utils import get_graphql_content
 
 WAREHOUSE_CREATE_MUTATION = """
@@ -16,20 +17,6 @@ mutation createWarehouse($input: WarehouseCreateInput!) {
   }
 }
 """
-
-
-DEFAULT_ADDRESS = {
-    "firstName": "John Saleor",
-    "lastName": "Doe Mirumee",
-    "companyName": "Saleor Commerce",
-    "streetAddress1": "	14208 Hawthorne Blvd",
-    "streetAddress2": "",
-    "postalCode": "90250",
-    "country": "US",
-    "city": "Hawthorne",
-    "countryArea": "CA",
-    "phone": "+12025550163",
-}
 
 
 def create_warehouse(
